fix(payment): PAYMENTS-7407 rename PPSDK failed files to failure to match updated API
